# Speed-up rand: Tausworthe RNG with shared random state.