--- a/.buildkite/pipeline.yml
+++ b/.buildkite/pipeline.yml
@@ -57,7 +57,7 @@ steps:
   - label: "CUDA 11.2"
     plugins:
       - JuliaCI/julia#v1:
-          version: 1.6
+          version: 1.6-nightly
       - JuliaCI/julia-test#v1:
           test_args: "--thorough"
       - JuliaCI/julia-coverage#v1:
@@ -79,7 +79,7 @@ steps:
   - label: "CUDA 11.1"
     plugins:
       - JuliaCI/julia#v1:
-          version: 1.6
+          version: 1.6-nightly
       - JuliaCI/julia-test#v1:
           test_args: "--thorough"
       - JuliaCI/julia-coverage#v1:
@@ -101,7 +101,7 @@ steps:
   - label: "CUDA 11.0"
     plugins:
       - JuliaCI/julia#v1:
-          version: 1.6
+          version: 1.6-nightly
       - JuliaCI/julia-test#v1:
           test_args: "--thorough"
       - JuliaCI/julia-coverage#v1:
@@ -123,7 +123,7 @@ steps:
   - label: "CUDA 10.2"
     plugins:
       - JuliaCI/julia#v1:
-          version: 1.6
+          version: 1.6-nightly
       - JuliaCI/julia-test#v1:
           test_args: "--thorough"
       - JuliaCI/julia-coverage#v1:
@@ -145,7 +145,7 @@ steps:
   - label: "CUDA 10.1"
     plugins:
       - JuliaCI/julia#v1:
-          version: 1.6
+          version: 1.6-nightly
       - JuliaCI/julia-test#v1:
           test_args: "--thorough"
       - JuliaCI/julia-coverage#v1: