--- a/src/device/intrinsics.jl
+++ b/src/device/intrinsics.jl
@@ -41,7 +41,6 @@ include("intrinsics/memory_dynamic.jl")
 include("intrinsics/atomics.jl")
 include("intrinsics/misc.jl")
 include("intrinsics/wmma.jl")
-include("intrinsics/random.jl")
 
 # functionality from libdevice
 #