Continuation of https://github.com/JuliaGPU/CUDA.jl/pull/772, inspired by https://forums.developer.nvidia.com/t/random-numbers-inside-the-kernel/14222/10. This uses a Combined Tausworthe generator that reads global state from shared memory. Seeds are fed from the CPU RNG when the CuModule is loaded, and are stored in constant memory (alongside with other RNG params).

Performance is great:

```
julia> @benchmark CUDA.@sync(broadcast!(()->rand(Float32), a)) setup=(a=CuArray{Float32}(undef, 1024))
BenchmarkTools.Trial: 
  memory estimate:  176 bytes
  allocs estimate:  5
  --------------
  minimum time:     5.408 μs (0.00% GC)
  median time:      5.717 μs (0.00% GC)
  mean time:        5.721 μs (0.00% GC)
  maximum time:     11.573 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     6

julia> @benchmark CUDA.@sync(broadcast!(()->rand(Float32), a)) setup=(a=CuArray{Float32}(undef, 1024, 1024))
BenchmarkTools.Trial: 
  memory estimate:  176 bytes
  allocs estimate:  5
  --------------
  minimum time:     25.503 μs (0.00% GC)
  median time:      26.029 μs (0.00% GC)
  mean time:        26.179 μs (0.00% GC)
  maximum time:     45.937 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> @benchmark CUDA.@sync(broadcast!(()->rand(Float32), a)) setup=(a=CuArray{Float32}(undef, 1024, 1024, 1024))
BenchmarkTools.Trial: 
  memory estimate:  176 bytes
  allocs estimate:  5
  --------------
  minimum time:     24.563 ms (0.00% GC)
  median time:      24.968 ms (0.00% GC)
  mean time:        25.033 ms (0.00% GC)
  maximum time:     25.364 ms (0.00% GC)
  --------------
  samples:          194
  evals/sample:     1
```

Copying from https://github.com/JuliaGPU/CUDA.jl/pull/772:

> CURAND:
> 
> ```
> julia> @benchmark CUDA.@sync(rand!(a)) setup=(a=CuArray{Float32}(undef, 1024))
> BenchmarkTools.Trial: 
>   memory estimate:  32 bytes
>   allocs estimate:  2
>   --------------
>   minimum time:     4.539 μs (0.00% GC)
>   median time:      4.694 μs (0.00% GC)
>   mean time:        4.700 μs (0.00% GC)
>   maximum time:     9.496 μs (0.00% GC)
>   --------------
>   samples:          10000
>   evals/sample:     7
> 
> julia> @benchmark CUDA.@sync(rand!(a)) setup=(a=CuArray{Float32}(undef, 1024, 1024))
> BenchmarkTools.Trial: 
>   memory estimate:  32 bytes
>   allocs estimate:  2
>   --------------
>   minimum time:     15.074 μs (0.00% GC)
>   median time:      15.798 μs (0.00% GC)
>   mean time:        16.095 μs (0.00% GC)
>   maximum time:     35.990 μs (0.00% GC)
>   --------------
>   samples:          10000
>   evals/sample:     1
> 
> julia> @benchmark CUDA.@sync(rand!(a)) setup=(a=CuArray{Float32}(undef, 1024, 1024, 1024))
> BenchmarkTools.Trial: 
>   memory estimate:  192 bytes
>   allocs estimate:  6
>   --------------
>   minimum time:     19.067 ms (0.00% GC)
>   median time:      104.840 ms (0.33% GC)
>   mean time:        99.918 ms (0.33% GC)
>   maximum time:     105.639 ms (0.32% GC)
>   --------------
>   samples:          51
>   evals/sample:     1
> ```
> 
> CUDA.jl:
> 
> ```
> julia> @benchmark CUDA.@sync(broadcast!(()->rand(Float32), a)) setup=(a=CuArray{Float32}(undef, 1024))
> BenchmarkTools.Trial: 
>   memory estimate:  208 bytes
>   allocs estimate:  7
>   --------------
>   minimum time:     8.752 μs (0.00% GC)
>   median time:      8.967 μs (0.00% GC)
>   mean time:        8.990 μs (0.00% GC)
>   maximum time:     31.119 μs (0.00% GC)
>   --------------
>   samples:          10000
>   evals/sample:     3
> 
> julia> @benchmark CUDA.@sync(broadcast!(()->rand(Float32), a)) setup=(a=CuArray{Float32}(undef, 1024, 1024))
> BenchmarkTools.Trial: 
>   memory estimate:  208 bytes
>   allocs estimate:  7
>   --------------
>   minimum time:     433.062 μs (0.00% GC)
>   median time:      452.577 μs (0.00% GC)
>   mean time:        452.522 μs (0.00% GC)
>   maximum time:     568.405 μs (0.00% GC)
>   --------------
>   samples:          9469
>   evals/sample:     1
> 
> julia> @benchmark CUDA.@sync(broadcast!(()->rand(Float32), a)) setup=(a=CuArray{Float32}(undef, 1024, 1024, 1024))
> ERROR: Out of GPU memory trying to allocate 4.000 GiB
> 
> julia> sum(sizeof, map(kernel->kernel.random_state, values(CUDA.cufunction_cache[device()]))) |> Base.format_bytes
> "4.004 GiB"
> ```

So a 15x speed-up, getting close to CURAND, and no restrictions wrt. launch sizes anymore.

Still TODO: figure out whether we want to mix in the block ID, as identical threads will currently yield the same random numbers across blocks:

```julia
function kernel()
    @cuprintln("thread $(threadIdx().x): $(rand())")
    return
end

@cuda threads=2 blocks=2 kernel()
```
```
thread 1: 0.605515
thread 2: 0.707695
thread 1: 0.605515
thread 2: 0.707695
```

EDIT: fixed that; even though I don't really know what I'm doing, the results look superficially OK:

```
julia> quantile(vec(Array(a)), [0.0, 0.25, 0.5, 0.75, 1.0])
5-element Vector{Float64}:
 2.384185791015625e-7
 0.24934574961662292
 0.49834924936294556
 0.7480988800525665
 0.9999868869781494
```

cc @simsurace
cc @S-D-R: this will now benefit a lot from https://github.com/JuliaGPU/CUDA.jl/pull/552, in case you'd want to incorporate that in your work.