@simsurace 2021-03-26
Using this in my binomial kernel improves its speed quite a bit (30-40%), see [this comment](https://github.com/JuliaGPU/BinomialGPU.jl/pull/9#issuecomment-808283457), moving them closer to the ones I'm getting in Julia 1.5.4. There is no way to use these new RNGs in Julia 1.5 to cross-check, is there?