@simsurace 2021-03-26
> I expected greater improvements though, as you can see in my results, but you're probably not generating many random numbers (or the overhead is in other operations you're performing).

BTRS generates 2-3 uniforms on average, and yes, the other operations are not negligible. That's why I found that up to count=17 it's still better to use the naive algorithm, which needs count number of uniform RVs. But 30-40% improvement is not bad at all.