@simsurace 2021-04-01
> The output looks OK, as judged by `quantile`, or by looking for duplicates (`rand(Float32, 1024, 1024)` generates about 5% duplicates, which is similar to Base's RNG. With Float64 it's 0.01%). I don't suppose this would pass BigCrush though, and it's annoying to set-up and use (the Julia package assumes a scalar `rand()`, for one).

What is the duplicate test you are using?