@simsurace 2021-04-01
> ```
> function nonunique(x::AbstractArray{T}) where T
>     uniqueset = Set{T}()
>     duplicatedset = Set{T}()
>     for i in x
>         if(i in uniqueset)
>             push!(duplicatedset, i)
>         else
>             push!(uniqueset, i)
>         end
>     end
>     collect(duplicatedset)
> end
> 
> @show 100 * length(nonunique(vec(A))) / length(A)
> ```

A quick calculation shows that for 2^20 samples of `Float32`s there should be an estimated `2^32*(1-(1-1/2^32)^(2^20)) = 1048448` unique values, which means that just 128 should be duplicates. This is much less than 5%, and becomes even less if you don't count duplicates of duplicates (which `nonunique` above doesn't do). Or am I making an error here?