@simsurace 2021-04-02
> shrug I just compared against Base's RNG, which produces similar results:
> 
> ```
> julia> A = rand(Float32, 1024, 1024);
> 
> julia> Adups = nonunique(A);
> 
> julia> @show 100 * length(nonunique(vec(A))) / length(A)
> (100 * length(nonunique(vec(A)))) / length(A) = 5.735969543457031
> 5.735969543457031
> ```
> 
> But that doesn't matter, as the crush failures are worrisome.

I think these are too many collisions by two-three orders of magnitude, see e.g. [this article](https://arxiv.org/pdf/2003.08009.pdf) which confirms the estimate I made above. So I should probably file an issue for Base as well.