@simsurace 2021-04-05
> shrug I just compared against Base's RNG, which produces similar results:
> 
> ```
> julia> A = rand(Float32, 1024, 1024);
> 
> julia> Adups = nonunique(A);
> 
> julia> @show 100 * length(nonunique(vec(A))) / length(A)
> (100 * length(nonunique(vec(A)))) / length(A) = 5.735969543457031
> 5.735969543457031
> ```
> 
> But that doesn't matter, as the crush failures are worrisome.

That's at least two orders of magnitude more collisions than expected, see [this paper](https://arxiv.org/pdf/2003.08009.pdf), which confirms the estimate I made earlier. This should probably be filed as an issue for Base as well?