@simsurace 2021-04-05
Actually the number of duplicates is as expected, as confirmed in [issue #40355 here](https://github.com/JuliaLang/julia/issues/40355). There are actually just `2^23`, not `2^32`, elements to choose from, and so the formula above gives an answer which is in line with the number of duplicates seen both in Base.rand and also this RNG here.